Preparing for Android release. Had to rename package because Google Play complained about existing app with the same package.
