+++ android/app/build.gradle
@@ -44,7 +44,7 @@ android {
 
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
-        applicationId "com.joinseeds.wallet"
+        applicationId "com.joinseeds.seedswallet"
         minSdkVersion 18
         targetSdkVersion 29
         versionCode flutterVersionCode.toInteger()
