+++ pubspec.yaml
@@ -1,7 +1,7 @@
 name: seeds
 description: SEEDS Wallet
 
-version: 1.1.0+4
+version: 1.0.1+12
 
 environment:
   sdk: ">=2.6.0 <3.0.0"
